Combine search and notification services
