--- search-service/Dockerfile
@@ -1,11 +1,0 @@
-FROM node:20-alpine
-
-WORKDIR /app
-
-COPY package*.json ./
-
-RUN npm install
-
-COPY . .
-
-CMD ["npm", "start"]
